8233787: Break cycle in vm_version* includes Reviewed-by: kbarrett, mdoerr
