+++ src/hotspot/cpu/sparc/vm_version_sparc.hpp
@@ -25,8 +25,8 @@
 #ifndef CPU_SPARC_VM_VERSION_SPARC_HPP
 #define CPU_SPARC_VM_VERSION_SPARC_HPP
 
+#include "runtime/abstract_vm_version.hpp"
 #include "runtime/globals_extension.hpp"
-#include "runtime/vm_version.hpp"
 
 class VM_Version: public Abstract_VM_Version {
   friend class VMStructs;
