+++ src/hotspot/cpu/zero/register_zero.hpp
@@ -27,7 +27,7 @@
 #define CPU_ZERO_REGISTER_ZERO_HPP
 
 #include "asm/register.hpp"
-#include "vm_version_zero.hpp"
+#include "runtime/vm_version.hpp"
 
 class VMRegImpl;
 typedef VMRegImpl* VMReg;
