+++ src/hotspot/cpu/zero/vm_version_zero.hpp
@@ -26,8 +26,8 @@
 #ifndef CPU_ZERO_VM_VERSION_ZERO_HPP
 #define CPU_ZERO_VM_VERSION_ZERO_HPP
 
+#include "runtime/abstract_vm_version.hpp"
 #include "runtime/globals_extension.hpp"
-#include "runtime/vm_version.hpp"
 
 class VM_Version : public Abstract_VM_Version {
  public:
