+++ src/hotspot/os/windows/os_perf_windows.cpp
@@ -31,7 +31,7 @@
 #include "runtime/os_perf.hpp"
 #include "runtime/os.hpp"
 #include "utilities/macros.hpp"
-#include "vm_version_ext_x86.hpp"
+#include CPU_HEADER(vm_version_ext)
 #include <math.h>
 #include <psapi.h>
 #include <TlHelp32.h>
